LI-8 | Move top bar composables to LimboTopBar.kt file
