+++ app/src/main/java/com/app/limboapp/screens/HomeScreen.kt
@@ -38,6 +38,7 @@ import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.app.limboapp.R
+import com.app.limboapp.common.LimboLogoWithPointsAndProfile
 import com.app.limboapp.model.ChapterData
 import com.app.limboapp.model.Person
 import com.app.limboapp.nav.LimboBottomNavigation
