+++ app/src/main/java/com/app/limboapp/screens/ReedemPointsScreen.kt
@@ -24,6 +24,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.app.limboapp.R
+import com.app.limboapp.common.LimboLogoWithPointsAndProfile
 import com.app.limboapp.nav.LimboBottomNavigation
 import com.app.limboapp.ui.theme.BlackBackground
 import com.app.limboapp.ui.theme.Montserrat
